+++ internal/metadata/metadata.go
@@ -36,8 +36,6 @@ type Metadata struct {
 	DOTTYStatus        AgentStatus `json:"dotty_status,omitempty"`
 	SSHInfo            *SSHInfo    `json:"ssh_info,omitempty"`
 	ManagedKeysEnabled *bool       `json:"managed_keys_enabled,omitempty"`
-	// TroubleshootingAgent contains the configuration for the troubleshooting agent on a Droplet
-	TroubleshootingAgent *TroubleshootingAgent `json:"troubleshooting_agent,omitempty"`
 }
 
 // SSHInfo contains the information of the sshd service running on the droplet
@@ -46,17 +44,4 @@ type SSHInfo struct {
 	Port uint16 `json:"port,omitempty"`
 	// HostKeys is the public ssh keys of the droplet, needed for identifying the droplet
 	HostKeys []string `json:"host_keys,omitempty"`
-}
-
-// TroubleshootingAgent represents the configuration for the troubleshooting agent on a Droplet.
-type TroubleshootingAgent struct {
-	// InvestigationUUID is the UUID of the investigation currently being
-	// performed by the agent.
-	InvestigationUUID string `json:"investigation_uuid,omitempty"`
-	// TriggeredAt is an ISO 8601 timestamp indicating when the alert associated
-	// with the investigation was triggered.
-	TriggeredAt string `json:"triggered_at,omitempty"`
-	// Requesting is a list of artifacts that the agent is being requested to
-	// collect, e.g. "file:/var/log/syslog"
-	Requesting []string `json:"requesting,omitempty"`
 }